--- src/types/DiaryDetail.ts
@@ -1,5 +1,0 @@
-export type DiaryDetail = {
-  title: string;
-  name: string;
-  description: string;
-};
